buildkitd: allow custom worker platforms for containerd
